Move ApplicationInfosPage
