+++ packages/core/admin/admin/src/pages/SettingsPage/pages/ApplicationInfosPage/index.js
@@ -5,7 +5,7 @@ import { HeaderLayout, Layout, ContentLayout } from '@strapi/parts/Layout';
 import { Main } from '@strapi/parts/Main';
 import { Box } from '@strapi/parts/Box';
 import { Grid, GridItem } from '@strapi/parts/Grid';
-import { H3 , Text, TableLabel } from '@strapi/parts/Text';
+import { H3, Text, TableLabel } from '@strapi/parts/Text';
 
 import { Stack } from '@strapi/parts/Stack';
 import { Link } from '@strapi/parts/Link';
